Improve SEO, AEO and UX foundations across core pages
