+++ components/home/hero.tsx
@@ -1,4 +1,4 @@
-'use client'
+﻿'use client'
 
 import { Button } from "@/components/ui/button"
 import Image from "next/image"
@@ -14,12 +14,8 @@ export default function Hero() {
         <div className="flex flex-col items-center gap-7 lg:gap-8">
           <div className="grid md:grid-cols-3 gap-6 lg:gap-8 items-center w-full rounded-[32px] border border-[#f3d7d7] bg-white/85 px-5 py-6 shadow-[0_20px_60px_rgba(144,20,14,0.06)] backdrop-blur-sm sm:px-7 lg:px-10 lg:py-8">
             <div className="md:col-span-2 space-y-4 text-center md:text-center">
-              <h1 className="text-4xl lg:text-5xl xl:text-6xl text-gray-900 leading-tight tracking-tight font-contrail-one">
-                {t('hero.title')}
-              </h1>
-              <p className="text-base lg:text-lg text-gray-700 leading-relaxed font-arimo max-w-3xl mx-auto">
-                {t('hero.description')}
-              </p>
+              <h1 className="text-4xl lg:text-5xl xl:text-6xl text-gray-900 leading-tight tracking-tight font-contrail-one">{t("hero.title")}</h1>
+              <p className="text-base lg:text-lg text-gray-700 leading-relaxed font-arimo max-w-3xl mx-auto">{t("hero.description")}</p>
               <div className="w-40 sm:w-52 h-1.5 bg-gradient-to-r from-[#90140e] via-[#bc2222] to-[#f45e5e] rounded-full mx-auto" />
             </div>
 
@@ -31,9 +27,10 @@ export default function Hero() {
           <div className="w-full">
             <Image
               src="/images/DSC_0044.webp"
-              alt="Jóvenes colaborando en educación"
+              alt="Jovenes colaborando en educacion"
               width={1400}
               height={500}
+              sizes="(max-width: 640px) 100vw, (max-width: 1024px) 92vw, 1400px"
               className="w-full h-[260px] sm:h-[320px] lg:h-[420px] rounded-3xl sm:rounded-[52px] lg:rounded-[72px] shadow-[0_30px_80px_rgba(144,20,14,0.16)] object-cover"
               priority
             />
@@ -41,13 +38,13 @@ export default function Hero() {
 
           <div className="flex flex-col sm:flex-row gap-3 md:gap-4 items-center justify-center">
             <Button asChild className="bg-[#90140e] hover:bg-[#a01810] text-white px-8 py-5 rounded-full text-lg md:text-xl font-bold shadow-lg transition-transform transform hover:scale-105 w-[240px] lg:w-[260px] font-contrail">
-              <Link href="/contactanos">{t('hero.voluntariado')}</Link>
+              <Link href="/contactanos">{t("hero.voluntariado")}</Link>
             </Button>
             <Button asChild className="bg-[#bd2222] hover:bg-[#d12828] text-white px-8 py-5 rounded-full text-lg md:text-xl font-bold shadow-lg transition-transform transform hover:scale-105 w-[240px] lg:w-[260px] font-contrail">
-              <Link href="/contactanos">{t('hero.alianzas')}</Link>
+              <Link href="/contactanos">{t("hero.alianzas")}</Link>
             </Button>
             <Button asChild className="bg-[#f45e5e] hover:bg-[#f67a7a] text-white px-8 py-5 rounded-full text-lg md:text-xl font-bold shadow-lg transition-transform transform hover:scale-105 w-[240px] lg:w-[260px] font-contrail">
-              <Link href="/donar">{t('hero.donar')}</Link>
+              <Link href="/donar">{t("hero.donar")}</Link>
             </Button>
           </div>
         </div>
