+++ public/manifest.webmanifest
@@ -1,0 +1,22 @@
+﻿{
+  "name": "Fundacion Lideres de Ansenuza",
+  "short_name": "FLA",
+  "description": "Programas educativos y de liderazgo juvenil de Fundacion Lideres de Ansenuza.",
+  "start_url": "/",
+  "display": "standalone",
+  "background_color": "#ffffff",
+  "theme_color": "#bc2222",
+  "lang": "es-AR",
+  "icons": [
+    {
+      "src": "/images/Logos/fla-logo-blanco.webp",
+      "sizes": "192x192",
+      "type": "image/webp"
+    },
+    {
+      "src": "/images/Logos/fla-logo-blanco.webp",
+      "sizes": "512x512",
+      "type": "image/webp"
+    }
+  ]
+}
